commit: (백준 - Greedy) 1202 보석 도둑 (priceSum long에서 int로 변환. 정답 처리됨)

+++ src/main/java/com/example/algorithm/boj/Greedy/Boj1202.java
@@ -40,7 +40,7 @@ public class Boj1202 {
 
         PriorityQueue<Integer> queue = new PriorityQueue<>((p1, p2) -> Integer.compare(p2, p1)); // 가격 내림 차순 정렬
 
-        int priceSum = 0; // 보석 총 가격
+        long priceSum = 0; // 보석 총 가격
 
         int idx = 0; // 보석 idx
         for (int i = 0; i < k; i++) {
